refactor: move core parse-and-generate steps from sde-generate.js to new parseAndGenerate() function in compile package
